Untested run of collectstatic before every web app startup
I didn't realize I'd have to do this, and I feel bad about it, but it'll do
for now. Docker bind mounts only go from host to container, not the other way
around, so we need to copy files from the container to the bind mount to make
them visible on the host (so nginx can serve the static content).