Fix automatic schema generation
We use different serializer classes for different subclasses of Study, and
this behavior assumed that we could call StudyDetailView.get_object inside
StudyDetailView.get_serializer_class.

This is not the case when the OpenAPI schema is automatically generated; the
schema generation process tries to obtain the serializer class without any
object PK. Fix StudyDetailView.get_serializer_class to fall back to
StudySerializer if no PK is given.